Convert get parameters to a validated array; Use a static function for ILL book requests.
Anytime a whole set of inter-related variables are named, this is probably better done via
a hash (keyed array).  This allows for each of validation and for passing into functions.
Some functional methods of classes don't depend on the object state and can be defined
statically, so they can be called without an instance of the object.